remove comments, prepare for merge
