+++ listenbrainz_spark/year_in_music/top_discoveries.py
@@ -1,8 +1,10 @@
+from datetime import datetime, date, time
+
 from more_itertools import chunked
 
 from listenbrainz_spark.path import RECORDING_ARTIST_DATAFRAME
 from listenbrainz_spark.stats import run_query
-from listenbrainz_spark.year_in_music.utils import create_tracks_of_the_year
+from listenbrainz_spark.utils import get_listens_from_dump
 
 USERS_PER_MESSAGE = 1000
 MAX_ARTIST_OCCURRENCE = 2
@@ -61,3 +63,20 @@ def generate_top_discoveries(year):
         "year": year,
         "type": "year_in_music_playlists_end"
     }
+
+
+def create_tracks_of_the_year(year):
+    end = datetime.combine(date(year, 12, 31), time.max)
+    listens = get_listens_from_dump(end=end)
+    listens.createOrReplaceTempView("listens_for_tracks_of_year")
+    query = f"""
+            SELECT user_id
+                 , recording_mbid
+                 , count(*) AS score
+              FROM listens_for_tracks_of_year
+             WHERE recording_mbid IS NOT NULL
+          GROUP BY user_id
+                 , recording_mbid
+            HAVING date_part('YEAR', min(listened_at)) = {year}
+    """
+    run_query(query).createOrReplaceTempView("tracks_of_year")
